requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
redis==5.0.1
rq==1.15.1
scikit-learn==1.4.0
shap==0.44.1
numpy==1.26.3
pandas==2.2.0
pytesseract==0.3.10
Pillow
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
python-dotenv==1.0.0
pydantic>=2.0
pydantic-settings>=2.0
sentry-sdk[fastapi]==1.40.0
pytest==7.4.4
httpx==0.26.0